Extract remote_action code from sender
Currently, the only way to run a `remote_action` is to call it through
the `action-runner`.

This means even if there is code within the `action-runner` that wants
to call a local action it will need to schedule it via HTTP / UDS.

As a first step to making this more efficient, move the `remote_action`
code out to a directly callable fn.
